Moved builders into their own files and changed their name slightly
